requirements.txt
Flask==2.0.1
Flask-Security-Too==3.3.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5
Flask-WTF==0.15.1
pycryptodome==3.11.0
pytz==2023.3
Werkzeug==2.0.1
email_validator==1.1.3